+++ firmware/controllers/algo/shift_torque_reduction_controller.cpp
@@ -47,6 +47,7 @@ void ShiftTorqueReductionController::updateTriggerPinState() {
             updateTriggerPinState(
                 engineConfiguration->torqueReductionTriggerPin,
                 engineConfiguration->torqueReductionTriggerPinMode,
+                /*invertPhysicalPin*/false,
                 engine->engineState.lua.torqueReductionState
             );
             break;
@@ -55,6 +56,7 @@ void ShiftTorqueReductionController::updateTriggerPinState() {
             updateTriggerPinState(
                 engineConfiguration->launchActivatePin,
                 engineConfiguration->launchActivatePinMode,
+                /*invertPhysicalPin*/false,
                 false
             );
             break;
@@ -63,6 +65,7 @@ void ShiftTorqueReductionController::updateTriggerPinState() {
             updateTriggerPinState(
                 engineConfiguration->clutchDownPin,
                 engineConfiguration->clutchDownPinMode,
+                /*invertPhysicalPin*/false,
                 engine->engineState.lua.clutchDownState
             );
             break;
@@ -71,6 +74,7 @@ void ShiftTorqueReductionController::updateTriggerPinState() {
             updateTriggerPinState(
                 engineConfiguration->clutchUpPin,
                 engineConfiguration->clutchUpPinMode,
+                /*invertPhysicalPin*/true,
                 engine->engineState.lua.clutchUpState
             );
             break;
@@ -92,6 +96,7 @@ static bool isShiftTorqueBelowTemperatureThreshold() {
 void ShiftTorqueReductionController::updateTriggerPinState(
     const switch_input_pin_e pin,
     const pin_input_mode_e mode,
+    const bool invertPhysicalPin,
     const bool invalidPinState
 ) {
   if (!torqueReductionTriggerPinState) {
@@ -106,7 +111,7 @@ void ShiftTorqueReductionController::updateTriggerPinState(
     isTorqueReductionTriggerPinValid = isBrainPinValid(pin);
     const bool previousTorqueReductionTriggerPinState = torqueReductionTriggerPinState;
     if (isTorqueReductionTriggerPinValid) {
-        torqueReductionTriggerPinState = efiReadPin(pin, mode);
+        torqueReductionTriggerPinState = efiReadPin(pin, mode) ^ invertPhysicalPin;
     } else {
         torqueReductionTriggerPinState = invalidPinState;
     }