+++ firmware/controllers/algo/shift_torque_reduction_controller.h
@@ -16,7 +16,7 @@ public:
 
 private:
 	void updateTriggerPinState();
-	void updateTriggerPinState(switch_input_pin_e pin, pin_input_mode_e mode, bool invalidPinState);
+	void updateTriggerPinState(switch_input_pin_e pin, pin_input_mode_e mode, const bool invertPhysicalPin, bool invalidPinState);
 
 	void updateTimeConditionSatisfied();
 	void updateRpmConditionSatisfied();